Integrate parking simulation in splash
